+++ components/update_client/test_configurator.cc
@@ -57,7 +57,7 @@ TestConfigurator::TestConfigurator(PrefService* pref_service)
 
 TestConfigurator::~TestConfigurator() = default;
 
-int TestConfigurator::InitialDelay() const {
+double TestConfigurator::InitialDelay() const {
   return initial_time_;
 }
 
@@ -163,7 +163,7 @@ void TestConfigurator::SetOnDemandTime(int seconds) {
   ondemand_time_ = seconds;
 }
 
-void TestConfigurator::SetInitialDelay(int seconds) {
+void TestConfigurator::SetInitialDelay(double seconds) {
   initial_time_ = seconds;
 }
